fix initstate to make the state buffer usable in setstate
setstate could use the results of previous initstate or setstate
calls (they return the old state buffer), but the documentation
requires that an initialized state buffer should be possible to
use in setstate immediately, which means that initstate should
save the generator parameters in it.

I also removed the copyright notice since it is present in the
copyright file.